requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
telethon==1.34.0
aiogram>=3.0.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
asyncpg==0.29.0
redis==5.0.1
feedparser==6.0.10
openai==1.12.0
langchain==0.1.5
gitpython==3.1.41
tenacity==8.2.3
jinja2==3.1.3
httpx==0.26.0
twscrape==0.17.0
pytz==2024.1
jieba==0.42.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
akshare>=1.12.0
PyGithub>=2.1.0
pyyaml>=6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
respx>=0.21.0
aiosqlite>=0.19.0
